adds some functionality and rest api

+++ src/main/java/de/oth/tdanylenko/tdanbank/entity/Address.java
@@ -49,7 +49,7 @@ public class Address {
         this.streetAddition = streetAddition;
     }
 
-    public long getHouseNr() {
+    public int getHouseNr() {
         return houseNr;
     }
